+++ rustler/src/types/elixir_struct.rs
@@ -14,7 +14,7 @@ pub fn get_ex_struct_name(map: Term) -> NifResult<Atom> {
     let env = map.get_env();
     // In an Elixir struct the value in the __struct__ field is always an atom.
     map.map_get(atom::__struct__().to_term(env))
-        .and_then(|e| Atom::from_term(e))
+        .and_then(Atom::from_term)
 }
 
 pub fn make_ex_struct<'a>(env: Env<'a>, struct_module: &str) -> NifResult<Term<'a>> {