+++ rustler_codegen/src/map.rs
@@ -56,7 +56,7 @@ pub fn transcoder_decorator(ast: &syn::DeriveInput) -> TokenStream {
         #encoder
     };
 
-    gen.into()
+    gen
 }
 
 pub fn gen_decoder(
@@ -96,7 +96,7 @@ pub fn gen_decoder(
         }
     };
 
-    gen.into()
+    gen
 }
 
 pub fn gen_encoder(
@@ -134,5 +134,5 @@ pub fn gen_encoder(
         }
     };
 
-    gen.into()
+    gen
 }