+++ rustler_codegen/src/tuple.rs
@@ -37,7 +37,7 @@ pub fn transcoder_decorator(ast: &syn::DeriveInput) -> TokenStream {
         #encoder
     };
 
-    gen.into()
+    gen
 }
 
 pub fn gen_decoder(
@@ -88,7 +88,7 @@ pub fn gen_decoder(
         }
     };
 
-    gen.into()
+    gen
 }
 
 pub fn gen_encoder(
@@ -134,5 +134,5 @@ pub fn gen_encoder(
         }
     };
 
-    gen.into()
+    gen
 }